Fixes embedded flags
When an embedded flag is removed it now is unchecked. This doesn't
change when a flag is checked without using the embedded flag
ie: clicking a flag with the mouse will not be changed by embedded flags
However, an embedded flag will always force the checkbox to remain on
until it has been removed from the embedded flags, which will immediatly
uncheck it.

Process uses a set to track the previous embedded flags, then
does set difference on the previous flags and the current flags
to determine which flags have changed. These two lists can be easily
iterated over, the first set are flags to check, the second set are
flags to uncheck.

closes #10