Experimental CLI started
I have started the CLI which will be a traditional UI for the project.

I also moved the Qt gui to the interfaces module. From there it is
imported and run as normal.
The CLI currently has the ability to display the table generated from
the groups part of the controller.